+++ libs/cardano-ledger-binary/test/Main.hs
@@ -2,6 +2,7 @@ module Main where
 
 import System.IO (BufferMode (LineBuffering), hSetBuffering, hSetEncoding, stdout, utf8)
 import qualified Test.Cardano.Ledger.Binary.Failure as Failure
+import qualified Test.Cardano.Ledger.Binary.GoldenSpec as GoldenSpec
 import qualified Test.Cardano.Ledger.Binary.PlainSpec as PlainSpec
 import qualified Test.Cardano.Ledger.Binary.RoundTripSpec as RoundTripSpec
 import qualified Test.Cardano.Ledger.Binary.Success as Success
@@ -23,6 +24,7 @@ spec = do
   PlainSpec.spec
   describe "Versioned" $ do
     RoundTripSpec.spec
+    GoldenSpec.spec
     Failure.spec
     Success.spec